Rename Picker to BasePicker
