+++ src/components/Picker/index.native.js
@@ -1,5 +1,5 @@
 import React, {forwardRef} from 'react';
-import BasePicker from './Picker';
+import BasePicker from './BasePicker';
 
 export default forwardRef((props, ref) => (
     // eslint-disable-next-line react/jsx-props-no-spreading
